push 13/06

+++ laravel01/app/Http/Controllers/MahasiswaController.php
@@ -14,7 +14,13 @@ class MahasiswaController extends Controller
      */
     public function index()
     {
-        $mahasiswa = Mahasiswa::all();
+        if(auth()->user()->role == 'D') {
+            $mahasiswa = Mahasiswa::where('user_id', auth()->user()->id)->get();
+            //select * from mahasiswas where user_id = 1
+        } else {
+            $mahasiswa = Mahasiswa::all();
+        }
+
         return view('mahasiswa.index')
                 ->with('mahasiswa', $mahasiswa);
     }
@@ -35,6 +41,10 @@ class MahasiswaController extends Controller
     {
     
         // return $request;
+        if ($request->user()->cannot('create',
+        Mahasiswa::class)) {
+            abort(403);
+        }
 
         $val = $request->validate([
             'npm' => "required|max:10", 
@@ -87,6 +97,10 @@ class MahasiswaController extends Controller
     public function update(Request $request, Mahasiswa $mahasiswa)
     {
         // dd($request)
+        if (auth()->user->cannot('update', $mahasiswa)) {
+            abort(403);
+    }
+
         if($request->url_foto) { //jika ada file foto yang dilampirkan
             $val = $request->validate([
                 // 'npm' => "required|max:10", 
@@ -132,8 +146,13 @@ class MahasiswaController extends Controller
     public function destroy(Mahasiswa $mahasiswa)
     {
         // dd($mahasiswa);
+        if (auth()->user->cannot('delete', $mahasiswa)) {
+                abort(403);
+        }
         File::delete('foto/'. $mahasiswa['url_foto']);
         $mahasiswa->delete(); //hapus data mahasiswa
         return redirect()->route('mahasiswa.index')->with('success', 'Data berhasil dihapus.');
+
     }
+    
 }
